adjusted folder structure and added views folder
