--- config/database.config.js
@@ -1,3 +1,0 @@
-module.exports = {
-    url: `mongodb+srv://${process.env.DB_USERNAME}:${process.env.DB_PASSWORD}@cluster0.xmhee.mongodb.net/anyLease`
-}
