+++ middleware/auth.js
@@ -1,32 +1,25 @@
-const jwt = require("jsonwebtoken");
-require('dotenv').config();
-const { SECRET } = process.env;
+const jwt = require("express-jwt");
 
+const getTokenFromHeaders = (req) => {
+    const { headers: { authorization } } = req;
 
-module.exports = (req, res, next) => {
-    //Get token from header
-    const token = req.header("x-auth-token");
+    if (authorization && authorization.split(" ")[0] === "Bearer") {
+        return authorization.split(" ")[1]
+    }
 
-    //Check if token doesn't exist
-    if (!token)
-     return res
-     .status(401)
-     .json({
-         statusCode: 401,
-         message: "No token, authorisation denied!"
-     });
-
-     //else... token exists
-     try {
-         const decoded = jwt.verify(token, SECRET);
-
-         //Assign user to request object
-         req.user = decoded.user;
-         next();
-     } catch (err) {
-        res.status(401).json({
-            statusCode: 401,
-            message: "Token is not valid"
-        });
-     }
+    return null;
 }
+
+const jwtObject = {
+    secret: "secret",
+    userProperty: "payload",
+    getToken: getTokenFromHeaders,
+    algorithms: ['sha1', 'RS256', 'HS256'],
+};
+
+const auth = {
+    required: jwt(jwtObject),
+    optional: jwt({ ...jwtObject, credentialsRequired: false })
+}
+
+module.exports = auth
